modularization: extract XML generator
